test(integrity): merge specs around flows, cover the simplified shapes
Merge the six integrity API spec files into two flow-oriented specs
(core.api.spec.ts: store/relay/checker/renewal, admin.api.spec.ts:
admin routes + metadata historization) matching the refactored joint-
anchor shapes: single revision key per dataset (no class segment),
{status, breach?} check responses, synchronous enable/_fix, and
{operation, origin, reason?} revision context. Add revisionsPrefix()
to tests/support/integrity.ts.
